app/requirements.txt
python-dotenv
requests
pydantic[email]
streamlit
fastapi[standard]
google-generativeai
asyncio
beautifulsoup4
langchain
langchain-community
langchain-huggingface
langchain-text-splitters
qdrant-client
langchain-qdrant
rich
openai
redis
logfire[fastapi]
slowapi